Extract event data
